==> notebooks/04_sanity_check_analytics.ipynb <==
import pandas as pd

fact = pd.read_parquet("../data/analytics/hmda/fact_loans_2023.parquet")
fact.head()


import pandas as pd

fact = pd.read_parquet("../data/analytics/hmda/fact_loans_2023.parquet")

agg_purpose = (
    fact
    .groupby(["state_code", "income_bucket", "loan_purpose", "applicant_gender"])
    .agg(
        applications=("approved", "count"),
        approval_rate=("approved", "mean"),
        median_loan_amount=("loan_amount", "median")
    )
    .reset_index()
)

agg_purpose.head(10)
